Fix detection of "soft" portage restarts
Portage does not log anything when restarting itself due to
the --keep-going flag. As a result, we don't detect exits the normal
way, which makes it look like the package that caused the exit is still
compiling.

This change assumes that `emerge (1 of` is a good enough signal that we
should abandon open sessions. Naturally, this will break with multiple
portage instances running at the same time. Until Portage learns to log
a soft restart, it's the best we can do.