Switch to /proc as source of running compile
While in theory, the log file should give us the truth of what compiles
are running, in reality, sometimes golop in its old form would show
aborted/failed compiles as still going on. Typically, this was due to
some compile of foo failing while the long-running parallel compile of
bar was still going on. This would show both foo and bar still running.
This situation also sometimes happened when multiple independent emerges
(not just parallel compiles) were going on.

So instead of reading current state from the log file, we use /proc.
There are downsides to this approach:

- We now depend on *having* /proc (not necessarily true with prefix
  installations) and it being readable for the current user (some
  paranoid setups make /proc/[pid]/cmdline only readable by the
  process's user and root).
- This pretty ugly, since it uses a global variable. It could probably
  be worked around by returning and passing the latestStart variable
  around.
- It parses the log file on -c even if there are no compilations
  running. I had hoped that somewhere in the process tree above the
  sandbox, there would be a process that lives during the whole set of
  ebuild phases. Unfortunately, the only process for which that is true
  is the top-level emerge process, which is not useful in this case.

As long as Portage (or emerge.log) doesn't change fundamentally, this
probably the best we can do. The neatest solution overall would be the
ability to talk to the emerge process about what is currently doing.